Extract HttpStream
